+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   - The dataset selection step is now displayed before going to the proof step
   - Users receive a data request submitted email after submitting their data
     request
+- **Organization Changes**
+  - A new data request viewer role has been added to organizations
 
 ## 53.0.0 (May 31, 2021)